requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
psycopg2-binary
nltk